Add metabolic matrix
Former-commit-id: ef2ffc0987c01af1bcdf8dcce05bb8cb1c051418
Former-commit-id: 35fda27ae156a613661e4e5dcc3088bcf7ec09f1

+++ code/lammps1Feb2014/src/atom.h
@@ -67,8 +67,8 @@ class Atom : protected Pointers {
   //type
   char **typeName;
   double *ks, *growth, *yield;
-  int **catCoeff;
-  int **anabCoeff;
+  double **catCoeff;
+  double **anabCoeff;
 
   //nutrient
   int nNutrients;
